refactor: 로비 페이지 컴포넌트 나누기
