+++ client/src/components/Profile.tsx
@@ -1,0 +1,14 @@
+export default function Profile() {
+  const user = JSON.parse(localStorage.getItem('user') as string);
+
+  return (
+    <div className="mb-6 flex items-center justify-center gap-3">
+      <img
+        className="w-12 rounded-full"
+        src={user.avatar_url}
+        alt="프로필 이미지"
+      />
+      <p className="text-lg font-semibold">{user.username}</p>
+    </div>
+  );
+}
