+++ client/src/components/buttons/RoomCreateButton.tsx
@@ -1,0 +1,7 @@
+export default function RoomCreateButton() {
+  return (
+    <button className="flex h-[33px] w-[150px] items-center justify-center rounded-lg bg-green-500 font-medium text-white hover:bg-green-400">
+      Create room
+    </button>
+  );
+}
