+++ client/src/components/buttons/RoomJoinButton.tsx
@@ -1,0 +1,7 @@
+export default function RoomJoinButton() {
+  return (
+    <button className="h-[33px] w-[60px] items-center justify-center rounded-lg bg-green-500 font-medium text-white hover:bg-green-400">
+      Join
+    </button>
+  );
+}
